Entity Component System overhaul, Structural Changes, Improved EnemySpawnManager, Better Debugging Tools, Background
Lately, I've been trying to optimize code a lot more, especially on high number of entities being alive at the same time causing issues.

I've tried and focused on the topic where, calling virtual functions a billion million trillion sextillion times each frame (from UIManager and EntityManager). Spent days, on this.

But eventually, I've just dumped the idea in the end, and recovered back to where I was. I've seen many things/errors/issues/bugs/features/backgroundstuff/compileroptimizations and they helped me on C++, definitely.

So I got back home with the lessons I learnt and just cut the optimization on that field entirely, at least for now! I'll just try to implement features into the game, as it is more fun this way rather than completely locking myself into one topic only. Just have fun, playing around in the code, and we'll see what happens in the near future.

+++ Client/src/Miscellaneous/Random.h
@@ -12,7 +12,7 @@ private:
 
 	// Static member initialization function to seed the random number generator with the current time
 	static std::mt19937& GetRNG() {
-		static std::mt19937 rng(std::chrono::system_clock::now().time_since_epoch().count());
+		static std::mt19937 rng(static_cast<unsigned int>(std::chrono::system_clock::now().time_since_epoch().count()));
 		return rng;
 	}
